Refactored code.| #1821
